체단실 가려고 급 커밋

+++ block/shape_type/shape_type.h
@@ -5,14 +5,14 @@
 class ShapeType
 {
 public:
-  ShapeType(int shape_type = 0);
+  ShapeType(short int shape_type = 0);
 
   void rotateCw();
   void rotateAcw();
-  const int& getIndex() const;
+  const short int& getIndex() const;
 
 private:
-  int shape_type_;
+  short int shape_type_;
 };
 
 #endif
